nnlos/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
httpx==0.27.0

# Google APIs
google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# XLSX parsing (shipment files)
openpyxl==3.1.5

# Scheduling (background worker)
apscheduler==3.10.4

# Templating (web UI)
jinja2==3.1.4

# AI email drafting (Phase 5)
anthropic==0.34.2
